Add features: Docker image, Netlify serverless deployment, updating tests for API v0.2
