+++ .dockerignore
@@ -1,0 +1,6 @@
+node_modules
+npm-debug.log
+.git
+.DS_Store
+*.ts
+*.map
