+++ .gitignore
@@ -1,3 +1,5 @@
 .env
 js
 node_modules
+# Local Netlify folder
+.netlify
