+++ package.json
@@ -11,7 +11,7 @@
     "test": "npm run build && node --test js/tests/**/*.js"
   },
   "author": "Huynh Nhan Thap",
-  "license": "MIT",
+  "license": "See LICENSE",
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
@@ -25,6 +25,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.19.2"
+    "express": "^4.19.2",
+    "serverless-http": "^4.0.0"
   }
 }
